* Moving `mask_saturated` into image utils. * Fixing make_pretty_image for fz files
